Split release, build, and test steps, add caching
This both allows better flexibility in build and the ability to test
across multiple libc versions. It should also speed up builds somewhat,
and this is leveraged for the unit and E2E test matrices.

It also provides a couple convenient side effects:

- Test binaries (aside from Miri for obvious reasons) are uploaded as
  artifacts, allowing debugging of them locally using the built binary.
  (Due to size, these are retained for only a week.)
- Release binaries are uploaded as artifacts, allowing validation of
  them and even testing without needing to cut a release.

Also, had to revise the E2E script to accept a binary as a parameter.