Remove caching and merge lint steps
The goal is to reduce build time

A solid 2 minutes of the build is wasted on initializing the cache for
both Clippy and Rustfmt each, while it only takes a couple minutes to
fetch all dependencies in the first place. Also, it often likes to take
a minute or three just to schedule a single task.

Miri uses the nightly toolchain, so it wouldn't benefit from caching
installation from the stable toolchain.

All in all, it should go from this (7m metered time, 8-10m real time):

- 2 minutes to download dependencies for Clippy
- 2 minutes to initialize cache in Clippy
- 2 minutes to initialize cache in Rustfmt
- 0 minutes to read from cache in build
- 15 seconds each to install toolchain for Clippy, Rustfmt, and build
- 1-3 minutes each to start Clippy, Rustfmt, Shellcheck, and build

to this (4-5m metered time, 6-8m real time):

- 2 minutes to download dependencies for lint
- 2 minutes to download dependencies for build
- 15 seconds each to install toolchain for lint and build
- 1-3 minutes each to start lint and build